Moved all providers to Providers folder and removed holder classes
