+++ Extension/src/LanguageServer/Providers/documentSymbolProvider.ts
@@ -3,8 +3,8 @@
  * See 'LICENSE' in the project root for license information.
  * ------------------------------------------------------------------------------------------ */
 import * as vscode from 'vscode';
-import { DefaultClient, LocalizeDocumentSymbol, GetDocumentSymbolRequestParams, GetDocumentSymbolRequest } from './client';
-import * as util from '../common';
+import { DefaultClient, LocalizeDocumentSymbol, GetDocumentSymbolRequestParams, GetDocumentSymbolRequest } from '../client';
+import * as util from '../../common';
 
 
 export class DocumentSymbolProvider implements vscode.DocumentSymbolProvider {
